producer/requirements.txt
websocket-client==1.7.0
confluent-kafka==2.3.0